Check for the event of pods using chaosengine_controller
