+++ cmd/manager/main.go
@@ -113,9 +113,6 @@ func main() {
 		log.Info(err.Error())
 	}
 
-	// Strating event listener
-	go controller.EventListner()
-
 	log.Info("Starting the Cmd.")
 
 	// Start the Cmd
